Problem 3 commit 2
Predicate added to find the minimum of a list. If only one value exists in the list, then it is the minimum. Otherwise, if the head of the list is smaller than the most recently calculated minimum, it is the minimum of the list. Else, if the head of the list continues to be larger than the calculated minimum, the minimum is the smallest in the list.