Problem 3 commit 4
Predicate added to find the smallest value in L1 that is larger than the minimum of L2. First, the minimum of L2 is found by calling the findMin predicate. Next, a list of the values in L1 that are larger than the minimum of L2 is created using the findLarger predicate. Finally, the minimum of that list is found.